rename models to singular
